Portfolio 3
Added HockeyStatTrackerKernel, HockeyStatTracker, and components.jar.
